+++ src/Client/ApiClientInterface.php
@@ -11,12 +11,15 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
  *     "aymdev.request_id"?: string,
  *     "aymdev.cache_ttl"?: int,
  *     "aymdev.cache_expiration"?: \DateTimeInterface,
- *     "aymdev.cache_error_ttl"?: int
+ *     "aymdev.cache_error_ttl"?: int,
+ *     "aymdev.log_request_body"?: bool,
+ *     "aymdev.log_response_body"?: bool,
+ *     "aymdev.log_error_response_body"?: bool
  * }
  *
  * @phpstan-type ApiClientOptions mixed[]|array{
- *      user_data?: UserDataOptions
- *  }
+ *     user_data?: UserDataOptions
+ * }
  */
 interface ApiClientInterface extends HttpClientInterface
 {
@@ -43,4 +46,22 @@ interface ApiClientInterface extends HttpClientInterface
      * The value is the duration in seconds.
      */
     public const string CACHE_ERROR_DURATION = 'aymdev.cache_error_ttl';
+
+    /**
+     * Request option used to log the request body.
+     * Must be set to `true`.
+     */
+    public const string LOG_REQUEST_BODY = 'aymdev.log_request_body';
+
+    /**
+     * Request option used to log the response body.
+     * Must be set to `true`.
+     */
+    public const string LOG_RESPONSE_BODY = 'aymdev.log_response_body';
+
+    /**
+     * Request option used to log the response body only on error.
+     * Must be set to `true`.
+     */
+    public const string LOG_ERROR_RESPONSE_BODY = 'aymdev.log_error_response_body';
 }