+++ src/Log/RequestLogger.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpClient\Response\AsyncContext;
 
 /**
  * @internal
+ *
  * @phpstan-import-type ApiClientOptions from ApiClientInterface
  */
 class RequestLogger
@@ -37,7 +38,9 @@ class RequestLogger
         $responseOrContext = [
             'method' => $responseOrContext->getInfo('http_method'),
             'url' => $responseOrContext->getInfo('url'),
+            'request_body' => $this->getRequestBody($options),
             'response_status' => $status,
+            'response_body' => $this->getResponseBody($responseOrContext, $options),
             'time' => $duration,
             'cache' => $responseOrContext instanceof CachedResponse,
             'error' => $responseOrContext->getInfo('error'),
@@ -48,5 +51,52 @@ class RequestLogger
         } else {
             $this->logger->info($message, $responseOrContext);
         }
+    }
+
+    /**
+     * @param ApiClientOptions $options
+     * @throws \JsonException
+     */
+    private function getRequestBody(array $options): ?string
+    {
+        if (!($options['user_data'][ApiClientInterface::LOG_REQUEST_BODY] ?? false)) {
+            return null;
+        }
+
+        if (isset($options['json'])) {
+            return json_encode($options['json'], JSON_THROW_ON_ERROR);
+        }
+
+        // TODO: handle other types of request body
+        if (!isset($options['body']) || !is_string($options['body'])) {
+            return null;
+        }
+        return $options['body'];
+    }
+
+    /**
+     * @param ApiClientOptions $options
+     * @throws \JsonException
+     */
+    private function getResponseBody(CachedResponse|AsyncContext $responseOrContext, array $options): ?string
+    {
+        $logResponse = $options['user_data'][ApiClientInterface::LOG_RESPONSE_BODY] ?? false;
+        $logErrorResponse = $options['user_data'][ApiClientInterface::LOG_ERROR_RESPONSE_BODY] ?? false;
+        if (!$logResponse && !($logErrorResponse && $responseOrContext->getStatusCode() >= 300)) {
+            return null;
+        }
+
+        if ($responseOrContext instanceof AsyncContext) {
+            if (null === $stream = $responseOrContext->getContent()) {
+                return null;
+            }
+            return stream_get_contents($stream, offset: 0);
+        }
+
+        if ([] !== $jsonData = $responseOrContext->toArray(false)) {
+            return json_encode($jsonData, JSON_THROW_ON_ERROR);
+        }
+
+        return $responseOrContext->getContent(false);
     }
 }