+++ tests/Log/RequestLoggerTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Tests\AymDev\ApiClientBundle\Log;
 
 use AymDev\ApiClientBundle\Cache\CachedResponse;
+use AymDev\ApiClientBundle\Client\ApiClientInterface;
 use AymDev\ApiClientBundle\Log\RequestLogger;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
@@ -26,6 +27,15 @@ class RequestLoggerTest extends TestCase
         $duration = 42.42;
         $error = 'OK';
 
+        // Set request/response body that should not be logged
+        $request = [
+            'body' => 'content',
+            'json' => [
+                'key' => 'value',
+            ],
+        ];
+        $responseBody = 'response';
+
         $logger = self::createMock(LoggerInterface::class);
         $logger->expects(self::once())
             ->method($logMethod)
@@ -34,7 +44,9 @@ class RequestLoggerTest extends TestCase
                 [
                     'method' => $method,
                     'url' => $url,
+                    'request_body' => null,
                     'response_status' => $status,
+                    'response_body' => null,
                     'time' => $duration,
                     'cache' => false,
                     'error' => $error,
@@ -42,7 +54,7 @@ class RequestLoggerTest extends TestCase
             );
 
         $passthru = null;
-        $response = new MockResponse(info: [
+        $response = new MockResponse($responseBody, [
             'http_method' => $method,
             'url' => $url,
             'http_code' => $status,
@@ -52,7 +64,7 @@ class RequestLoggerTest extends TestCase
         $context = new AsyncContext($passthru, new MockHttpClient(), $response, $infos, null, 0);
 
         $requestLogger = new RequestLogger($logger);
-        $requestLogger->logRequest($duration, $context, []);
+        $requestLogger->logRequest($duration, $context, $request);
     }
 
     public static function provideResponses(): \Generator
@@ -74,5 +86,168 @@ class RequestLoggerTest extends TestCase
         $response = self::createMock(CachedResponse::class);
         $requestLogger = new RequestLogger($logger);
         $requestLogger->logRequest(1, $response, []);
+    }
+
+    public function testLogRequestBody(): void
+    {
+        $requestBody = 'content';
+        $request = [
+            'body' => $requestBody,
+            'user_data' => [
+                ApiClientInterface::LOG_REQUEST_BODY => true,
+            ],
+        ];
+
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['request_body'] === $requestBody),
+            );
+
+        $response = self::createMock(CachedResponse::class);
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $response, $request);
+    }
+
+    public function testLogJsonRequestBody(): void
+    {
+        $jsonBody = '{"key":"value"}';
+        $request = [
+            'json' => [
+                'key' => 'value'
+            ],
+            'user_data' => [
+                ApiClientInterface::LOG_REQUEST_BODY => true,
+            ],
+        ];
+
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['request_body'] === $jsonBody),
+            );
+
+        $response = self::createMock(CachedResponse::class);
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $response, $request);
+    }
+
+    public function testLogResponseBodyFromCachedResponse(): void
+    {
+        $responseBody = 'content';
+        $response = new CachedResponse(200, [], $responseBody, [], []);
+
+        $request = [
+            'user_data' => [
+                ApiClientInterface::LOG_RESPONSE_BODY => true,
+            ],
+        ];
+
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['response_body'] === $responseBody)
+            );
+
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $response, $request);
+    }
+
+    public function testLogResponseJsonBodyFromCachedResponse(): void
+    {
+        $jsonBody = '{"key":"value"}';
+        $jsonData = [
+            'key' => 'value',
+        ];
+        $response = new CachedResponse(200, [], '', $jsonData, []);
+
+        $request = [
+            'user_data' => [
+                ApiClientInterface::LOG_RESPONSE_BODY => true,
+            ],
+        ];
+
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['response_body'] === $jsonBody)
+            );
+
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $response, $request);
+    }
+
+    public function testLogResponseBodyFromContext(): void
+    {
+        $responseBody = 'content';
+        $request = [
+            'user_data' => [
+                ApiClientInterface::LOG_RESPONSE_BODY => true,
+            ],
+        ];
+
+        // Define context resource used by context
+        $resource = fopen('php://temp', 'w+');
+        if (false === $resource) {
+            throw new \RuntimeException('Unable to open temp file');
+        }
+        fwrite($resource, $responseBody);
+
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['response_body'] === $responseBody)
+            );
+
+        $passthru = null;
+        $response = new MockResponse($responseBody);
+        $infos = [];
+        $context = new AsyncContext($passthru, new MockHttpClient(), $response, $infos, $resource, 0);
+
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $context, $request);
+
+        fclose($resource);
+    }
+
+    #[DataProvider('provideLogErrorResponseBodyCases')]
+    public function testNotLogErrorResponseBodyWithoutError(int $statusCode, bool $expectsLog): void
+    {
+        $responseBody = 'content';
+        $response = new CachedResponse($statusCode, [], $responseBody, [], []);
+
+        $request = [
+            'user_data' => [
+                ApiClientInterface::LOG_ERROR_RESPONSE_BODY => true,
+            ],
+        ];
+
+        $expectedLog = $expectsLog ? $responseBody : null;
+        $logger = self::createMock(LoggerInterface::class);
+        $logger->expects(self::once())
+            ->method('info')
+            ->with(
+                self::isString(),
+                self::callback(fn(array $context) => $context['response_body'] === $expectedLog),
+            );
+
+        $requestLogger = new RequestLogger($logger);
+        $requestLogger->logRequest(1, $response, $request);
+    }
+
+    public static function provideLogErrorResponseBodyCases(): \Generator
+    {
+        yield [200, false];
+        yield [404, true];
     }
 }