+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - request logging
  - response caching
+ - optional body logging for request, response and error response
 
 [Unreleased]: https://github.com/AymDev/ApiClientBundle/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/AymDev/ApiClientBundle/releases/tag/v1.0.0